ver idk

+++ conv_sequential.c
@@ -1,12 +1,13 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <omp.h>
 
 int main(){
     // ---- input and malloc A, F ----
     int NA, NF;
     scanf("%d %d", &NA, &NF);
-    int *A = malloc(sizeof(int) * NA);
-    int *F = malloc(sizeof(int) * NF);
+    long long *A = malloc(sizeof(long long) * NA);
+    long long *F = malloc(sizeof(long long) * NF);
 
     for(int i = 0; i < NA; i++){
         scanf("%d", &A[i]);
@@ -18,17 +19,18 @@ int main(){
 
     // implement here
     int x = NA-NF+1;
-    int *result = malloc(sizeof(int) * x);
+    long long *result = malloc(sizeof(long long) * x);
     for(int i = 0; i < x; i++){
-        int res = 0;
+        long long res = 0;
         for(int j = 0; j < NF; j++){
             res += A[i+j] * F[NF-j-1];
         }
         result[i] = res;
     }
+    
 
     for(int i = 0; i < x; i++){
-        printf("%d\n", result[i]);
+        printf("%lld\n", result[i]);
     }
 
     // ---- free memory ----
